Moving HTTP wrapper classes to HTTP client file
